// src/phone/apps/wechat/dating/vnBgmManifest.generated.ts
export const VN_BGM_MANIFEST: { name: string; fileName: string; url: string }[] = [
  { name: "爱、双向奔赴、心动、感动", fileName: "爱、双向奔赴、心动、感动.mp3", url: `${import.meta.env.BASE_URL}BGM/%E7%88%B1%E3%80%81%E5%8F%8C%E5%90%91%E5%A5%94%E8%B5%B4%E3%80%81%E5%BF%83%E5%8A%A8%E3%80%81%E6%84%9F%E5%8A%A8.mp3` },
  { name: "暧昧场景 拉扯，亲密", fileName: "暧昧场景-拉扯，亲密.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%9A%A7%E6%98%A7%E5%9C%BA%E6%99%AF-%E6%8B%89%E6%89%AF%EF%BC%8C%E4%BA%B2%E5%AF%86.mp3` },
  { name: "安静、悲伤", fileName: "安静、悲伤.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%AE%89%E9%9D%99%E3%80%81%E6%82%B2%E4%BC%A4.mp3` },
  { name: "安静、海边、温馨、静谧", fileName: "安静、海边、温馨、静谧.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%AE%89%E9%9D%99%E3%80%81%E6%B5%B7%E8%BE%B9%E3%80%81%E6%B8%A9%E9%A6%A8%E3%80%81%E9%9D%99%E8%B0%A7.mp3` },
  { name: "安静、平静、遗憾", fileName: "安静、平静、遗憾.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%AE%89%E9%9D%99%E3%80%81%E5%B9%B3%E9%9D%99%E3%80%81%E9%81%97%E6%86%BE.mp3` },
  { name: "悲伤回忆、难过、低沉、堕落", fileName: "悲伤回忆、难过、低沉、堕落.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%82%B2%E4%BC%A4%E5%9B%9E%E5%BF%86%E3%80%81%E9%9A%BE%E8%BF%87%E3%80%81%E4%BD%8E%E6%B2%89%E3%80%81%E5%A0%95%E8%90%BD.mp3` },
  { name: "感动、心动、悲伤、遗憾、BE美学、凄美", fileName: "感动、心动、悲伤、遗憾、BE美学、凄美.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%84%9F%E5%8A%A8%E3%80%81%E5%BF%83%E5%8A%A8%E3%80%81%E6%82%B2%E4%BC%A4%E3%80%81%E9%81%97%E6%86%BE%E3%80%81BE%E7%BE%8E%E5%AD%A6%E3%80%81%E5%87%84%E7%BE%8E.mp3` },
  { name: "告白、恋人嬉戏、美好、暧昧、朦胧、日落", fileName: "告白、恋人嬉戏、美好、暧昧、朦胧、日落.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%91%8A%E7%99%BD%E3%80%81%E6%81%8B%E4%BA%BA%E5%AC%89%E6%88%8F%E3%80%81%E7%BE%8E%E5%A5%BD%E3%80%81%E6%9A%A7%E6%98%A7%E3%80%81%E6%9C%A6%E8%83%A7%E3%80%81%E6%97%A5%E8%90%BD.mp3` },
  { name: "告白表白、心动、悸动", fileName: "告白表白、心动、悸动.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%91%8A%E7%99%BD%E8%A1%A8%E7%99%BD%E3%80%81%E5%BF%83%E5%8A%A8%E3%80%81%E6%82%B8%E5%8A%A8.mp3` },
  { name: "欢快、开心1", fileName: "欢快、开心1.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%AC%A2%E5%BF%AB%E3%80%81%E5%BC%80%E5%BF%831.mp3` },
  { name: "欢快日常、轻松愉悦", fileName: "欢快日常、轻松愉悦.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%AC%A2%E5%BF%AB%E6%97%A5%E5%B8%B8%E3%80%81%E8%BD%BB%E6%9D%BE%E6%84%89%E6%82%A6.mp3` },
  { name: "回忆、纯爱、BE美学、凄美、宿命感、悲伤美化、", fileName: "回忆、纯爱、BE美学、凄美、宿命感、悲伤美化、.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%9B%9E%E5%BF%86%E3%80%81%E7%BA%AF%E7%88%B1%E3%80%81BE%E7%BE%8E%E5%AD%A6%E3%80%81%E5%87%84%E7%BE%8E%E3%80%81%E5%AE%BF%E5%91%BD%E6%84%9F%E3%80%81%E6%82%B2%E4%BC%A4%E7%BE%8E%E5%8C%96%E3%80%81.mp3` },
  { name: "回忆、凄美、感动、心动", fileName: "回忆、凄美、感动、心动.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%9B%9E%E5%BF%86%E3%80%81%E5%87%84%E7%BE%8E%E3%80%81%E6%84%9F%E5%8A%A8%E3%80%81%E5%BF%83%E5%8A%A8.mp3` },
  { name: "回忆、遗憾、意难平", fileName: "回忆、遗憾、意难平.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%9B%9E%E5%BF%86%E3%80%81%E9%81%97%E6%86%BE%E3%80%81%E6%84%8F%E9%9A%BE%E5%B9%B3.mp3` },
  { name: "回忆遗憾、安静、平静、释怀", fileName: "回忆遗憾、安静、平静、释怀.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%9B%9E%E5%BF%86%E9%81%97%E6%86%BE%E3%80%81%E5%AE%89%E9%9D%99%E3%80%81%E5%B9%B3%E9%9D%99%E3%80%81%E9%87%8A%E6%80%80.mp3` },
  { name: "日常、欢快", fileName: "日常、欢快.mp3", url: `${import.meta.env.BASE_URL}BGM/%E6%97%A5%E5%B8%B8%E3%80%81%E6%AC%A2%E5%BF%AB.mp3` },
  { name: "伤感、平淡", fileName: "伤感、平淡.mp3", url: `${import.meta.env.BASE_URL}BGM/%E4%BC%A4%E6%84%9F%E3%80%81%E5%B9%B3%E6%B7%A1.mp3` },
  { name: "失恋、回忆、悲伤、心碎、难过", fileName: "失恋、回忆、悲伤、心碎、难过.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%A4%B1%E6%81%8B%E3%80%81%E5%9B%9E%E5%BF%86%E3%80%81%E6%82%B2%E4%BC%A4%E3%80%81%E5%BF%83%E7%A2%8E%E3%80%81%E9%9A%BE%E8%BF%87.mp3` },
  { name: "失恋、意难平、遗憾、悲伤、难过", fileName: "失恋、意难平、遗憾、悲伤、难过.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%A4%B1%E6%81%8B%E3%80%81%E6%84%8F%E9%9A%BE%E5%B9%B3%E3%80%81%E9%81%97%E6%86%BE%E3%80%81%E6%82%B2%E4%BC%A4%E3%80%81%E9%9A%BE%E8%BF%87.mp3` },
  { name: "甜蜜、欢快、心动、热恋", fileName: "甜蜜、欢快、心动、热恋.mp3", url: `${import.meta.env.BASE_URL}BGM/%E7%94%9C%E8%9C%9C%E3%80%81%E6%AC%A2%E5%BF%AB%E3%80%81%E5%BF%83%E5%8A%A8%E3%80%81%E7%83%AD%E6%81%8B.mp3` },
  { name: "偷偷摸摸、搞怪、搞笑", fileName: "偷偷摸摸、搞怪、搞笑.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%81%B7%E5%81%B7%E6%91%B8%E6%91%B8%E3%80%81%E6%90%9E%E6%80%AA%E3%80%81%E6%90%9E%E7%AC%91.mp3` },
  { name: "乌龙、搞笑、无语", fileName: "乌龙、搞笑、无语.mp3", url: `${import.meta.env.BASE_URL}BGM/%E4%B9%8C%E9%BE%99%E3%80%81%E6%90%9E%E7%AC%91%E3%80%81%E6%97%A0%E8%AF%AD.mp3` },
  { name: "心动、暧昧、升温氛围感、emo、悲伤、亲密戏、吻", fileName: "心动、暧昧、升温氛围感、emo、悲伤、亲密戏、吻.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%BF%83%E5%8A%A8%E3%80%81%E6%9A%A7%E6%98%A7%E3%80%81%E5%8D%87%E6%B8%A9%E6%B0%9B%E5%9B%B4%E6%84%9F%E3%80%81emo%E3%80%81%E6%82%B2%E4%BC%A4%E3%80%81%E4%BA%B2%E5%AF%86%E6%88%8F%E3%80%81%E5%90%BB.mp3` },
  { name: "心动、轻松、愉悦", fileName: "心动、轻松、愉悦.mp3", url: `${import.meta.env.BASE_URL}BGM/%E5%BF%83%E5%8A%A8%E3%80%81%E8%BD%BB%E6%9D%BE%E3%80%81%E6%84%89%E6%82%A6.mp3` },
]